+++ django/applications/catmaid/static/js/widgets/overlay.js
@@ -3969,9 +3969,10 @@ var SkeletonAnnotations = {};
       let mirrorIndex = CATMAID.TracingOverlay.Settings.session.read_only_mirror_index;
       let api = self.api;
       if (!api && mirrorIndex > -1) {
-        api = CATMAID.Client.Settings.session.remote_servers[mirrorIndex - 1];
-        if (api) {
+        let remoteServer = CATMAID.Client.Settings.session.remote_servers[mirrorIndex - 1];
+        if (remoteServer) {
           dedicatedActiveSkeletonUpdate = true;
+          api = CATMAID.API.fromSetting(remoteServer);
         }
       }